refactor: CD-122 auto fill keyphrase and apply thunk when generating keyphrase
