+++ client/src/components/web-extension/Common/CreatePasswordPage/CreatePasswordPage.scss
