+++ client/src/components/web-extension/Common/CreatePasswordPage/index.jsx
