+++ client/src/components/web-extension/ImportWallet/ImportWallet.jsx
@@ -1,9 +1,7 @@
 import React from "react";
 import { useSelector } from 'react-redux';
-
 import { selectCreateWalletCurrentStep } from "@cd/selectors/createWallet";
-import CreatePasswordPage from "@cd/common/CreatePasswordPage";
-
+import CreatePasswordPage from '@cd/web-extension/Common/CreatePasswordPage';
 import ImportWalletPage from "./ImportWalletPage";
 
 import './ImportWallet.scss';
