+++ client/src/components/web-extension/ImportWallet/ImportWalletPage.jsx
@@ -4,10 +4,8 @@ import { Button, Form as FormBS } from 'react-bootstrap';
 import { useSelector } from 'react-redux';
 import { Formik, Form } from 'formik';
 import { KeyFactory } from 'casper-storage';
-
 import { selectCreateWalletTotalKeywords } from "@cd/selectors/createWallet";
 import { setNextStep, updateKeyphrase } from "@cd/actions/createWalletActions";
-
 import FieldKeyphrase from "./FieldKeyphrase";
 
 const ImportWallet = () => {
@@ -22,7 +20,7 @@ const ImportWallet = () => {
 		 * Make sure keyphrase has 12 words
 		 */
 		if (keyphrase.some(key => key === '')) {
-			errors.keyphrase = 'Keypharse does not empty';
+			errors.keyphrase = 'Keyphrase required.';
 
 			return errors;
 		}
@@ -31,7 +29,7 @@ const ImportWallet = () => {
 		 * Make sure keyphrase is valid with `KeyFactory.getInstance().validate(keyphrase)`
 		 */
 		if (!KeyFactory.getInstance().validate(keyphrase.join(' '))) {
-			errors.keyphrase = 'Keypharse is not validated';
+			errors.keyphrase = 'Keyphrase is invalid. Please recheck your keyphrase.';
 
 			return errors;
 		}
@@ -54,7 +52,7 @@ const ImportWallet = () => {
 			validate={onValidate}
 			onSubmit={onSubmit}
 		>
-			{({ errors, touched, setFieldValue, handleBlur, handleSubmit }) => {
+			{({ errors, touched, handleSubmit }) => {
 				return (
 					<Form noValidate onSubmit={handleSubmit} className="cd_we_create-wallet-layout--root">
 						<div className="cd_we_create-wallet-layout--body cd_we_create-keyphrase--box">
